Conflicts:
	plugins/pgpointcloud/test/PgpointcloudWriterTest.cpp

+++ include/pdal/PointContext.hpp
@@ -253,7 +253,7 @@ private:
         for (auto ui = used.begin(); ui != used.end(); ++ui)
         {
             m_dims->m_detail[*ui].m_offset = offset;
-            offset += m_dims->m_detail[*ui].size();
+            offset += (int)m_dims->m_detail[*ui].size();
         }
         m_ptBuf->setPointSize((size_t)offset);
     }
